Refactor: decompose draksetup commands into separate files
