+++ drakrun/drakrun/draksetup/__main__.py
@@ -1,0 +1,3 @@
+from .main import main
+
+main()
